pyproject.toml
[tool.poetry]
name = "mkdocs-extra-sass-plugin"
description = "This plugin adds stylesheets to your mkdocs site from `Sass`/`SCSS`."
version = "0.1.0"
authors = ["orzih <orzih@mail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/orzih/mkdocs-extra-sass-plugin"
keywords = ["mkdocs", "css", "sass", "scss"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.plugins."mkdocs.plugins"]
extra-sass = 'mkdocs_extra_sass_plugin.plugin:ExtraSassPlugin'

[tool.poetry.dependencies]
python = ">=3.6"
mkdocs = ">=1.1"
beautifulsoup4 = ">=4.6.3"
libsass = ">=0.15"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
flake8-bugbear = "^20.11.1"
flake8-isort = "^4.0.0"
flake8-print = "^4.0.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"
